@@ -12,30 +12,29 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.15",
+    "@angular/animations": "^8.0.1",
     "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
-    "@angular/http": "^7.2.15",
+    "@angular/common": "^8.0.1",
+    "@angular/compiler": "^8.0.1",
+    "@angular/core": "^8.0.1",
+    "@angular/forms": "^8.0.1",
     "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/router": "^7.2.15",
+    "@angular/platform-browser": "^8.0.1",
+    "@angular/platform-browser-dynamic": "^8.0.1",
+    "@angular/router": "^8.0.1",
     "angular-file": "^0.5.4",
     "angular-oauth2-oidc": "^5.0.2",
     "core-js": "^2.5.4",
     "d3": "^5.6.0",
     "d3-ng2-service": "^2.2.0",
     "file-saver": "^2.0.0",
-    "graph-topology": "1.1.1",
-    "graph-topology-model-lib": "1.0.5",
+    "graph-topology": "1.1.2",
+    "graph-topology-model-lib": "1.0.6",
     "hammerjs": "^2.0.8",
-    "kypo2-markdown-editor": "1.0.4",
+    "kypo2-markdown-editor": "1.0.2",
     "kypo2-trainings-hurdling-viz-lib": "0.9.1",
     "kypo2-trainings-visualization-overview-lib": "0.9.5",
-    "kypo2-user-and-group-management": "1.0.8",
+    "kypo2-user-and-group-management": "1.0.9",
     "marked": "^0.6.1",
     "moment": "^2.22.2",
     "ng-pick-datetime": "^6.0.10",
@@ -43,19 +42,19 @@
     "ngx-hotjar": "0.0.4",
     "ngx-markdown": " ^7.1.4",
     "ngx-papaparse": "^3.0.0",
-    "rxjs": "^6.3.3",
+    "rxjs": "^6.5.2",
     "typescript-collections": "^1.3.2",
-    "zone.js": "^0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.9",
-    "@angular/cli": "^7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.2.15",
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular/cli": "^8.0.3",
+    "@angular/compiler-cli": "^8.0.1",
+    "@angular/language-service": "^8.0.1",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~1.7.1",
@@ -66,6 +65,6 @@
     "protractor": "~5.3.0",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "~3.1.3"
+    "typescript": "~3.4.5"
   }
 }
