@@ -30,8 +30,8 @@ export class TraineeTrainingsTableComponent implements OnInit {
   isInErrorState = false;
   now = Date.now();
 
-  @ViewChild(MatPaginator) paginator: MatPaginator;
-  @ViewChild(MatSort) sort: MatSort;
+  @ViewChild(MatPaginator, { static: true }) paginator: MatPaginator;
+  @ViewChild(MatSort, { static: true }) sort: MatSort;
 
   constructor(
     private router: Router,
