@@ -9,27 +9,27 @@ import {AdminGuard} from './services/guards/admin-guard.service';
 const routes: Routes = [
   {
     path: 'home',
-    loadChildren: 'app/components/portal/portal.module#PortalModule',
+    loadChildren: () => import('app/components/portal/portal.module').then(m => m.PortalModule),
     canActivate: [AuthGuard]
   },
   {
     path: 'designer',
-    loadChildren: 'app/components/designer/designer-overview/designer-overview.module#DesignerOverviewModule',
+    loadChildren: () => import('app/components/designer/designer-overview/designer-overview.module').then(m => m.DesignerOverviewModule),
     canActivate: [DesignerGuard]
   },
   {
     path: 'organizer',
-    loadChildren: 'app/components/organizer/organizer-overview/organizer-overview.module#OrganizerOverviewModule',
+    loadChildren: () => import('app/components/organizer/organizer-overview/organizer-overview.module').then(m => m.OrganizerOverviewModule),
     canActivate: [OrganizerGuard]
   },
   {
     path: 'trainee',
-    loadChildren: 'app/components/trainee/trainee-overview/trainee-overview.module#TraineeOverviewModule',
+    loadChildren: () => import('app/components/trainee/trainee-overview/trainee-overview.module').then(m => m.TraineeOverviewModule),
     canActivate: [TraineeGuard]
   },
   {
     path: 'admin',
-    loadChildren: 'app/components/admin/admin.module#AdminModule',
+    loadChildren: () => import('app/components/admin/admin.module').then(m => m.AdminModule),
     canActivate: [AdminGuard]
   },
   {
